Dev masha

+++ ClientMonitor.Infrastructure.Database/Repositories/CpuRepository.cs
@@ -23,14 +23,17 @@ namespace ClientMonitor.Infrastructure.Database.Repositories
             var mon = new EntitiesCpu
             {
                 DateTime = info.DateTime,
-                BusyCpu= info.BusyCpu,
-                FreeCpu=info.FreeCpu,
+                BusyCpu = info.BusyCpu,
+                FreeCpu = info.FreeCpu,
             };
 
             db.ECpus.Add(mon);
 
             DateTime threeday = DateTime.Now.AddDays(-3);
-            db.ECpus.RemoveRange(db.ECpus.Where(x => x.DateTime < threeday));
+            if (db.ECpus.Any())
+            {
+                db.ECpus.RemoveRange(db.ECpus.Where(x => x.DateTime < threeday));
+            }
             db.SaveChanges();
         }
 
@@ -44,9 +47,15 @@ namespace ClientMonitor.Infrastructure.Database.Repositories
             DateTime start = new DateTime(DateTime.Now.Year, DateTime.Now.Month, DateTime.Now.Day, 6, 0, 0);
             DateTime average = new DateTime(DateTime.Now.Year, DateTime.Now.Month, DateTime.Now.Day, 15, 30, 0);
             DateTime end = average.AddDays(-1);
-            if (dateTime.Hour == 6)
+            List<string> cpus = new();
+            if (db.EDataForEdit.Any())
             {
-                List<string> cpus = new();
+                start = Convert.ToDateTime(db.EDataForEdit.Where(c => c.Name == "TimeFirst").Select(x => x.Value).FirstOrDefault());
+                average = Convert.ToDateTime(db.EDataForEdit.Where(c => c.Name == "TimeSecond").Select(x => x.Value).FirstOrDefault());
+            }
+
+            if (dateTime.Hour == start.Hour)
+            {
                 cpus.Add(Math.Round(db.ECpus.Where(p => p.DateTime > end && p.DateTime < start).Min(u => u.BusyCpu), 3).ToString());
                 double maxcpu = db.ECpus.Where(p => p.DateTime > end && p.DateTime < start).Max(u => u.BusyCpu);
                 var dtcpu = db.ECpus.FirstOrDefault(p => p.BusyCpu == maxcpu);
@@ -57,7 +66,6 @@ namespace ClientMonitor.Infrastructure.Database.Repositories
             }
             else
             {
-                List<string> cpus = new();
                 cpus.Add(Math.Round(db.ECpus.Where(p => p.DateTime > start && p.DateTime < average).Min(u => u.BusyCpu), 3).ToString());
                 double maxcpu = db.ECpus.Where(p => p.DateTime > start && p.DateTime < average).Max(u => u.BusyCpu);
                 var dtcpu = db.ECpus.FirstOrDefault(p => p.BusyCpu == maxcpu);
